Fix browser-safe audio helper imports
