+++ apps/web/src/lib/server/audio/sentence-tokens.ts
@@ -8,13 +8,11 @@
  */
 import { asc, eq } from 'drizzle-orm';
 
+import type { SentenceTokens } from '../../audio/sentence-tokens.js';
 import { db, schema } from '../db/index.js';
 import type { TextToken } from '../db/schema.js';
-
-export type SentenceTokens = {
-  sentenceIdx: number;
-  tokens: Array<{ id: string; surface: string; isWord: boolean }>;
-};
+export { interpolateSentenceMarks } from '../../audio/sentence-tokens.js';
+export type { SentenceMark, SentenceTokens } from '../../audio/sentence-tokens.js';
 
 export async function loadSentenceTokensForAudio(
   audioFileId: string,
@@ -50,45 +48,6 @@ export async function loadSentenceTokensForAudio(
       out.push(cur);
     }
     cur.tokens.push({ id: t.id, surface: t.surface, isWord: t.isWord });
-  }
-  return out;
-}
-
-export type SentenceMark = {
-  sentenceIdx: number;
-  startMs: number;
-  endMs: number;
-};
-
-/**
- * Linearly interpolate per-token timing across each sentence's
- * marked (startMs, endMs) range. Each word in a sentence gets an
- * equal slice; non-word tokens (punctuation / whitespace) are
- * skipped — the alignment table is per-token but the highlighter
- * only paints `isWord` spans anyway, so emitting rows for
- * punctuation just bloats the table.
- *
- * Pure function — unit-tested without the DB.
- */
-export function interpolateSentenceMarks(
-  sentences: SentenceTokens[],
-  marks: SentenceMark[],
-): Array<{ tokenId: string; startMs: number; endMs: number }> {
-  const marksBySentence = new Map<number, SentenceMark>();
-  for (const m of marks) marksBySentence.set(m.sentenceIdx, m);
-  const out: Array<{ tokenId: string; startMs: number; endMs: number }> = [];
-  for (const s of sentences) {
-    const mark = marksBySentence.get(s.sentenceIdx);
-    if (!mark) continue;
-    const wordTokens = s.tokens.filter((t) => t.isWord);
-    if (wordTokens.length === 0) continue;
-    const span = Math.max(0, mark.endMs - mark.startMs);
-    const slice = span / wordTokens.length;
-    for (let i = 0; i < wordTokens.length; i++) {
-      const start = Math.round(mark.startMs + slice * i);
-      const end = Math.round(mark.startMs + slice * (i + 1));
-      out.push({ tokenId: wordTokens[i]!.id, startMs: start, endMs: end });
-    }
   }
   return out;
 }
